+++ packages/neuron-wallet/src/utils/i18n.ts
@@ -1,14 +1,16 @@
 import i18n from 'i18next'
 import zh from 'locales/zh'
 import en from 'locales/en'
+import app from 'app'
 
 if (!i18n.isInitialized) {
+  const fallbackLng = ['zh', 'zh-CN'].includes(app.getLocale()) ? 'zh' : 'en'
   i18n.init({
     resources: {
       en,
       zh,
     },
-    fallbackLng: 'en',
+    fallbackLng,
     interpolation: {
       escapeValue: false,
     },